refactor: CardController 응답시 List 보내줄 수 있도록 수정
- sample data에 cardIndex가 status별로 0번만 있게 수정

closes #36

+++ todo-list/src/main/resources/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE CARD (
 
 CREATE TABLE LOG
 (
-    id                  int AUTO_INCREMENT PRIMARY KEY,
+    id                  BIGINT AUTO_INCREMENT PRIMARY KEY,
     title               varchar(50) NOT NULL,
     current_status      ENUM('TODO', 'PROGRESS', 'DONE'),
     prev_status         ENUM('TODO', 'PROGRESS', 'DONE') default NULL,
